+++ events/__init__.py
@@ -1,4 +1,4 @@
-from finance.events.Event import Event
+from finance.events.Condition import Condition
 from finance.events.PastEvent import PastEvent
 from finance.events.EventFinder import EventFinder
 from finance.events.MultipleEvents import MultipleEvents