+++ examples/Basic/EventStudy_Finder.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from finance.events import SampleEvents
+from finance.events import SampleConditions
 from finance.events import EventFinder
 
 # from finance.utils import DataAccess
@@ -9,7 +9,7 @@ evtf = EventFinder()
 evtf.symbols = ['AMD', 'CBG', 'AAPL']
 evtf.start_date = datetime(2008, 1, 1)
 evtf.end_date = datetime(2010, 12, 31)
-evtf.event = SampleEvents.went_below(10)
+evtf.condition = SampleConditions.went_below(10)
 evtf.search(useCache=False)
 
 print(evtf.num_events)